Fixed namespaces (moved interfaces to Interfaces, entities to Entity, exceptions to Exception.
