+++ src/Extension/Exception/NotMappedException.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace OAT\DependencyResolver\Extension;
+namespace OAT\DependencyResolver\Extension\Exception;
 
 class NotMappedException extends \Exception
 {
