+++ src/FileSystem/Exception/FileAccessException.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace OAT\DependencyResolver\FileSystem;
+namespace OAT\DependencyResolver\FileSystem\Exception;
 
 class FileAccessException extends \Exception
 {
